step counter and battery percentage

+++ include/BluetoothSpeakerApp.h
@@ -8,8 +8,8 @@ struct BluetoothSpeakerApp: public App
   BluetoothSpeakerApp(App *next_app): App(next_app) {}
 
   void setup();
-  void on_touch_down(uint16_t x, uint16_t y);
-  void on_touch_up(uint16_t x, uint16_t y);
+  void on_touch_down(uint32_t x, uint32_t y);
+  void on_touch_up(uint32_t x, uint32_t y);
   void on_button_up();
   void on_button_long_press();
   void update();
