+++ include/PingPongApp.h
@@ -9,8 +9,8 @@ struct PingPongApp: public App
   PingPongApp(App *next_app): App(next_app) {}
 
   void setup();
-  void on_touch_down(uint16_t x, uint16_t y);
-  void on_touch_up(uint16_t x, uint16_t y);
+  void on_touch_down(uint32_t x, uint32_t y);
+  void on_touch_up(uint32_t x, uint32_t y);
   void on_button_up();
   void on_button_long_press();
   void update();
